+++ app/src/main/java/com/example/kunbaapp/data/models/dto/NodeDto.kt
@@ -1,9 +1,9 @@
 package com.example.kunbaapp.data.models.dto
 
 data class NodeDto(
-    val nodeId : ULong,
-    val rootId : ULong,
-    val familyId : ULong,
+    val nodeId : Int,
+    val rootId : Int,
+    val familyId : Int,
     val firstName : String,
     val lastName : String,
     val gender : Char,
